Move G-code parser into src/parser and normalize comments
Relocates gcode-parser.ts (and its tests) into a parser/ directory to
make room for the upcoming slicer metadata parsers, and updates all
import paths plus the coverage threshold key.

parseCommand now trims the extracted comment once at the source, so
every consumer sees '; layer 1' and ';layer 1' identically. Slicers
emit a space after the semicolon, and downstream comment consumers
(notably the upcoming metadata parsers, which anchor patterns with ^)
rely on this normalization.
